+++ CMakeLists.txt
@@ -12,9 +12,11 @@ include_directories(${LLVM_INCLUDE_DIRS})
 separate_arguments(LLVM_DEFINITIONS_LIST NATIVE_COMMAND ${LLVM_DEFINITIONS})
 add_definitions(${LLVM_DEFINITIONS_LIST})
 
-add_executable(cornpiler main.cpp compiler_frontend.cpp compile.cpp parse_cli.cpp translate.cpp)
+add_executable(cornpiler main.cpp compiler_frontend.cpp compile.cpp parse_cli.cpp translate.cpp logger.cpp)
 
 add_subdirectory(ast)
+
+add_subdirectory(pre_ast)
 
 execute_process(COMMAND llvm-config-12 --system-libs --link-shared --libs all
                 OUTPUT_VARIABLE llvm_libs
