+++ TODO
@@ -2,3 +2,5 @@ Make every test case run correctly
 Split off the lower AST into its own folder, make every AST node inhabit its own .cpp file
 Make inline assembly work
 Proper error handling
+Move first level AST functions to their own .cpp files
+Remove the need for with_ classes and solve that with interfaces
