+++ pre_ast/CMakeLists.txt
@@ -1,0 +1,5 @@
+cmake_policy(SET CMP0076 NEW)
+target_sources(cornpiler 
+PUBLIC 
+ast.cpp
+)
