+++ pre_ast/with/args.hpp
@@ -1,0 +1,9 @@
+#pragma once
+#include "../ast.hpp"
+
+namespace ast_types {
+class with_args : virtual public AST {
+ public:
+  ast_body args;
+};
+}  // namespace ast_types
