+++ pre_ast/with/second_body.hpp
@@ -1,0 +1,9 @@
+#pragma once
+#include "../ast.hpp"
+
+namespace ast_types {
+class with_second_body : virtual public AST {
+ public:
+  ast_body second_body;
+};
+}  // namespace ast_types
