+++ pre_ast/with/type.hpp
@@ -1,0 +1,9 @@
+#pragma once
+#include "../ast.hpp"
+
+namespace ast_types {
+class with_type : virtual public AST {
+ public:
+  ast_body type;
+};
+}  // namespace ast_types
